refactor: 상품 전체 선택하는 버튼 분리
