+++ src/components/Cart/CartItemContainer/CartItemController/CartItemController.tsx
@@ -1,22 +1,11 @@
 /** @jsxImportSource @emotion/react */
-import { useRecoilState } from "recoil";
-import { ControllerStyle } from "./CartItemController.style";
-import { checkAllItemState } from "../../../../store/selector/selectors";
-import Checkbox from "../../../Button/Checkbox/Checkbox";
+
+import CheckAllButton from "./CheckAllButton/CheckAllButton";
 
 const CartItemController = () => {
-  const [state, setState] = useRecoilState(checkAllItemState);
-
-  const handleClick = () => {
-    setState((prev) => {
-      return !prev;
-    });
-  };
-
   return (
-    <div css={ControllerStyle}>
-      <Checkbox isCheck={state} onClick={handleClick} />
-      <div>전체 선택</div>
+    <div>
+      <CheckAllButton />
     </div>
   );
 };
