+++ src/components/Cart/CartItemContainer/CartItemController/CheckAllButton/CheckAllButton.style.ts
@@ -1,0 +1,9 @@
+import { css } from "@emotion/react";
+
+export const CheckAllButtonStyle = css({
+  display: "flex",
+  justifyContent: "flex-start",
+  alignItems: "center",
+
+  gap: "8px",
+});
